pyproject.toml
[project]
name = "mini-split-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Nilo Aires Jr.",email = "nilo_aires@hotmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.17.0,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "pytz (>=2026.1.post1,<2027.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "engine.settings"
python_files = ["tests*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 120
target-version = "py313"

exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "node_modules",
    "__pycache__",
    "migrations",
    "static",
    "media",
]
[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
